requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pyyaml==6.0.1
yfinance==0.2.37
pandas==2.2.1
aiofiles==23.2.1
httpx==0.27.0